Refactor ohlc api
